add configurations

+++ lms-admin/src/main/java/com/swe/lms/admin/api/service/impl/BookCategoryServiceImpl.java
@@ -2,17 +2,16 @@ package com.swe.lms.admin.api.service.impl;
 
 import com.swe.lms.admin.api.adapter.BookCategoryAdapter;
 import com.swe.lms.admin.api.dto.BookCategoryDTO;
-import com.swe.lms.admin.api.dto.BookDTO;
+import com.swe.lms.admin.api.model.Book;
 import com.swe.lms.admin.api.model.BookCategory;
 import com.swe.lms.admin.api.repository.BookCategoryRepository;
+import com.swe.lms.admin.api.repository.BookRepository;
 import com.swe.lms.admin.api.service.IBookCategoryService;
 import com.swe.lms.admin.api.util.ResponseUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.PostMapping;
-
-import javax.annotation.PostConstruct;
+import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
@@ -21,6 +20,10 @@ public class BookCategoryServiceImpl implements IBookCategoryService {
 
     @Autowired
     private BookCategoryRepository bookCategoryRepository;
+
+    @Autowired
+    private BookRepository bookRepository;
+
     @Override
     public ResponseEntity<?> insertBookCategory(BookCategoryDTO bookCategoryDTO) {
         BookCategory bookCategory = BookCategoryAdapter.convertToBookCategory(bookCategoryDTO);
@@ -44,6 +47,10 @@ public class BookCategoryServiceImpl implements IBookCategoryService {
     public ResponseEntity<?> deleteBookCategory(int bookCategoryId) {
         Optional<BookCategory> optBookCategory = bookCategoryRepository.findById(bookCategoryId);
         if (optBookCategory.isPresent()) {
+            List<Book> books = bookRepository.findByBookCategoryId(bookCategoryId);
+            if (null != books && books.size() > 0) {
+                return ResponseUtil.createBadRequest("Book Category is not empty.");
+            }
             bookCategoryRepository.deleteById(bookCategoryId);
             return ResponseUtil.createOK(optBookCategory.get(), String.format("BookCategory ID [%d]", bookCategoryId));
         }
